+++ crates/prelude/src/callback.rs
@@ -20,7 +20,7 @@ macro_rules! define {
     (#[$m:meta] $n:ident $(, $x:ident)*) => {
         #[$m] #[no_mangle]
         extern "C" fn $n (
-            ptr: extern "C" fn(*const u8 $(, usize ${ignore(x)})*),
+            ptr: extern "C" fn(*const u8 $(, usize ${ignore($x)})*),
             this: *const u8 $(, $x: usize)*
         ) {
             ptr(this $(, $x)*);